requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
cached-property==2.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.1
executing==2.2.1
fastembed==0.7.4
filelock==3.24.0
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
idna==3.11
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
jiter==0.13.0
jupyter_client==8.8.0
jupyter_core==5.9.1
loguru==0.7.3
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==2.2.6
onnxruntime==1.23.2
openai==2.20.0
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.7.0
portkey-ai==2.1.0
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.3.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.23.1
typer-slim==0.23.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
